Start moving pub structs to separate file
