+++ res/blif/adder.blif
@@ -1,4 +1,4 @@
-# Generated by Yosys 0.61+112 (git sha1 ac427a79b-dirty, x86_64-w64-mingw32-g++ 13.2.1 -O3)
+# Generated by Yosys 0.33 (git sha1 2584903a060)
 
 .model full_add
 .inputs a b cin
